move engine fixture to conftest.py
