+++ tests/conftest.py
@@ -1,0 +1,12 @@
+import pytest
+from sqlalchemy import Engine, create_engine
+
+
+@pytest.fixture(scope="module")
+def engine() -> Engine:
+    user = "postgres"
+    password = "postgres"
+    host = "127.0.0.1"
+    port = 5432
+    db_name = "postgres"
+    return create_engine(f"postgresql+psycopg://{user}:{password}@{host}:{port}/{db_name}")
