+++ tests/it/test_role.py
@@ -3,7 +3,7 @@ from datetime import datetime, timedelta
 import pytest
 from hypothesis import HealthCheck, given, settings
 from hypothesis import strategies as st
-from sqlalchemy import Engine, create_engine
+from sqlalchemy import Engine
 
 from postgres_declare.base_entity import Entity
 from postgres_declare.cluster_entities import Role
@@ -15,16 +15,6 @@ def test_role() -> YieldFixture[Role]:
     yield Role(name="test_role")
     Entity.entities = []
     Entity.check_if_any_exist = False
-
-
-@pytest.fixture
-def engine() -> Engine:
-    user = "postgres"
-    password = "postgres"
-    host = "127.0.0.1"
-    port = 5432
-    db_name = "postgres"
-    return create_engine(f"postgresql+psycopg://{user}:{password}@{host}:{port}/{db_name}")
 
 
 def test_does_not_exist(test_role: Role, engine: Engine) -> None:
